+++ context/app/static/js/components/Detail/visualization/VisualizationShareButton/constants.js
@@ -1,0 +1,6 @@
+const DEFAULT_LONG_URL_WARNING =
+  'Warning: this is a long URL which may be incompatible or load slowly with some browsers.';
+
+const DEFAULT_EMAIL_MESSAGE = 'Here is an interesting dataset I found in the HuBMAP Data Portal:';
+
+export { DEFAULT_LONG_URL_WARNING, DEFAULT_EMAIL_MESSAGE };
